Update imports for moved module
