+++ regions/shapes/rectangle.py
@@ -256,7 +256,7 @@ class RectanglePixelRegion(PixelRegion):
         ``selector.set_active(True)`` or ``selector.set_active(False)``.
         """
         from matplotlib.widgets import RectangleSelector
-        from .utils import MPL_VERSION
+        from .._utils.optional_deps import MPL_VERSION
 
         if hasattr(self, '_mpl_selector'):
             raise AttributeError('Cannot attach more than one selector to a region.')
