+++ regions/shapes/tests/test_circle.py
@@ -12,9 +12,9 @@ import pytest
 
 from ...core import PixCoord, RegionMeta, RegionVisual
 from ...tests.helpers import make_simple_wcs
+from ..._utils.optional_deps import HAS_MATPLOTLIB  # noqa
 from ..circle import CirclePixelRegion, CircleSkyRegion
 from .test_common import BaseTestPixelRegion, BaseTestSkyRegion
-from ..utils import HAS_MATPLOTLIB  # noqa
 
 
 @pytest.fixture(scope='session', name='wcs')
