+++ regions/shapes/tests/test_line.py
@@ -13,9 +13,9 @@ from astropy.wcs import WCS
 
 from ...core import PixCoord, RegionMeta, RegionVisual
 from ...tests.helpers import make_simple_wcs
+from ..._utils.optional_deps import HAS_MATPLOTLIB  # noqa
 from ..line import LinePixelRegion, LineSkyRegion
 from .test_common import BaseTestPixelRegion, BaseTestSkyRegion
-from ..utils import HAS_MATPLOTLIB  # noqa
 
 
 @pytest.fixture(scope='session', name='wcs')
